+++ cmd/lambda/accounts/list_test.go
@@ -1,173 +1,125 @@
 package main
 
 import (
-	"context"
-	"encoding/json"
-	"errors"
-	"net/http"
+	"fmt"
+	"io/ioutil"
+	"net/http/httptest"
 	"net/url"
-	"strings"
 	"testing"
 
-	"github.com/Optum/dce/pkg/api/response"
-	"github.com/Optum/dce/pkg/db"
-	"github.com/Optum/dce/pkg/db/mocks"
-	"github.com/aws/aws-lambda-go/events"
+	"github.com/Optum/dce/pkg/account"
+	"github.com/Optum/dce/pkg/account/accountiface/mocks"
+	"github.com/Optum/dce/pkg/config"
+	"github.com/gorilla/schema"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/mock"
 )
 
-func TestListController_Call(t *testing.T) {
+func TestGetAccounts(t *testing.T) {
 
-	t.Run("When the invoking Call and there are no errors", func(t *testing.T) {
-		expectedAccounts := createAccountsOutput()
-		mockAccountInput := createGetAccountsInput()
-		expectedAccountsResponse := &[]*response.AccountResponse{
-			{
-				ID:             "123456789",
-				AccountStatus:  "Ready",
-				LastModifiedOn: 1561149393,
-				CreatedOn:      1561149393,
-				AdminRoleArn:   "mock:role:arn",
-			},
-		}
-		mockDb := mocks.DBer{}
-		mockDb.On("GetAccounts", mockAccountInput).Return(*expectedAccounts, nil)
-		mockRequest := createGetSingleAccountRequest()
-
-		Dao = &mockDb
-
-		actualResponse, err := Handler(context.TODO(), mockRequest)
-		require.Nil(t, err)
-
-		parsedResponse := &[]*response.AccountResponse{}
-		err = json.Unmarshal([]byte(actualResponse.Body), parsedResponse)
-		require.Nil(t, err)
-
-		require.Equal(t, expectedAccountsResponse, parsedResponse, "it returns a list of accounts.")
-		require.Equal(t, actualResponse.StatusCode, 200, "it returns a 200.")
-	})
-
-	t.Run("When the query fails", func(t *testing.T) {
-		expectedError := errors.New("error")
-		mockAccountInput := createGetEmptyAccountsInput()
-		mockAccountOutput := createAccountsOutput()
-		mockDb := mocks.DBer{}
-		mockDb.On("GetAccounts", mockAccountInput).Return(*mockAccountOutput, expectedError)
-		mockRequest := events.APIGatewayProxyRequest{HTTPMethod: http.MethodGet, Path: "/accounts"}
-
-		Dao = &mockDb
-
-		actualResponse, err := Handler(context.TODO(), mockRequest)
-		require.Nil(t, err)
-
-		require.Equal(t, actualResponse.StatusCode, 500, "it returns a 500.")
-	})
-
-	t.Run("Building next URL", func(t *testing.T) {
-		body := strings.NewReader("")
-		request, err := http.NewRequest("GET", "http://example.com/api/accounts?limit=2", body)
-
-		assert.Nil(t, err)
-
-		nextParams := make(map[string]string)
-
-		nextParams["AccountId"] = "1"
-
-		nextURL := response.BuildNextURL(request, nextParams, url.URL{})
-
-		assert.Equal(t, url.Values{
-			"limit": {
-				"2",
-			},
-			"nextAccountId": {
-				"1",
-			},
-		}, nextURL.Query())
-
-	})
-
-	t.Run("Empty accounts", func(t *testing.T) {
-
-		accountsResult := createEmptyAccountsOutput()
-
-		mockAccountInput := createGetEmptyAccountsInput()
-		mockDb := mocks.DBer{}
-		mockDb.On("GetAccounts", mockAccountInput).Return(*accountsResult, nil)
-		mockRequest := createGetEmptyAccountsRequest()
-
-		Dao = &mockDb
-
-		actualResponse, err := Handler(context.Background(), mockRequest)
-		require.Nil(t, err)
-
-		parsedResponse := []response.AccountResponse{}
-		err = json.Unmarshal([]byte(actualResponse.Body), &parsedResponse)
-		require.Nil(t, err)
-
-		require.Equal(t, 0, len(parsedResponse), "empty result")
-		require.Equal(t, actualResponse.StatusCode, 200, "Returns a 200.")
-	})
-}
-
-func createGetEmptyAccountsInput() db.GetAccountsInput {
-	keys := make(map[string]string)
-	return db.GetAccountsInput{
-		StartKeys: keys,
+	type response struct {
+		StatusCode int
+		Body       string
 	}
-}
-
-func createGetAccountsInput() db.GetAccountsInput {
-	keys := make(map[string]string)
-	return db.GetAccountsInput{
-		StartKeys: keys,
-		AccountID: "987654321",
-		Status:    db.Ready,
-	}
-}
-
-func createGetSingleAccountRequest() events.APIGatewayProxyRequest {
-	q := make(map[string]string)
-	q[AccountIDParam] = "987654321"
-	q[StatusParam] = "ready"
-	return events.APIGatewayProxyRequest{
-		HTTPMethod:            http.MethodGet,
-		QueryStringParameters: q,
-		Path:                  "/accounts",
-	}
-}
-
-func createAccountsOutput() *db.GetAccountsOutput {
-	nextKeys := make(map[string]string)
-	accounts := []*db.Account{
+	tests := []struct {
+		name        string
+		expResp     response
+		expLink     string
+		query       *account.Account
+		retAccounts *account.Accounts
+		retErr      error
+		nextID      *string
+	}{
 		{
-			ID:             "123456789",
-			AccountStatus:  db.Ready,
-			LastModifiedOn: 1561149393,
-			CreatedOn:      1561149393,
-			AdminRoleArn:   "mock:role:arn",
+			name:  "get all accounts",
+			query: &account.Account{},
+			expResp: response{
+				StatusCode: 200,
+				Body:       "[]\n",
+			},
+			retAccounts: &account.Accounts{},
+			retErr:      nil,
+		},
+		{
+			name:  "get paged accounts",
+			query: &account.Account{},
+			expResp: response{
+				StatusCode: 200,
+				Body:       "[{\"id\":\"123456789012\"}]\n",
+			},
+			retAccounts: &account.Accounts{
+				account.Account{
+					ID: ptrString("123456789012"),
+				},
+			},
+			nextID:  ptrString("234567890123"),
+			expLink: "<https://example.com/unit/accounts?limit=1&nextId=234567890123>; rel=\"next\"",
+			retErr:  nil,
+		},
+		{
+			name: "fail to get accounts",
+			query: &account.Account{
+				ID: ptrString("abc123"),
+			},
+			expResp: response{
+				StatusCode: 500,
+				Body:       "{\"error\":{\"message\":\"unknown error\",\"code\":\"ServerError\"}}\n",
+			},
+			retAccounts: nil,
+			retErr:      fmt.Errorf("failure"),
 		},
 	}
-	return &db.GetAccountsOutput{
-		NextKeys: nextKeys,
-		Results:  accounts,
-	}
-}
 
-func createGetEmptyAccountsRequest() events.APIGatewayProxyRequest {
-	q := make(map[string]string)
-	return events.APIGatewayProxyRequest{
-		HTTPMethod:            http.MethodGet,
-		QueryStringParameters: q,
-		Path:                  "/accounts",
-	}
-}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			r := httptest.NewRequest("GET", "http://example.com/accounts", nil)
 
-func createEmptyAccountsOutput() *db.GetAccountsOutput {
-	nextKeys := make(map[string]string)
-	accounts := []*db.Account{}
-	return &db.GetAccountsOutput{
-		NextKeys: nextKeys,
-		Results:  accounts,
+			baseRequest = url.URL{}
+			baseRequest.Scheme = "https"
+			baseRequest.Host = "example.com"
+			baseRequest.Path = fmt.Sprintf("%s%s", "unit", "/accounts")
+
+			values := url.Values{}
+			err := schema.NewEncoder().Encode(tt.query, values)
+			assert.Nil(t, err)
+
+			r.URL.RawQuery = values.Encode()
+			w := httptest.NewRecorder()
+
+			cfgBldr := &config.ConfigurationBuilder{}
+			svcBldr := &config.ServiceBuilder{Config: cfgBldr}
+
+			accountSvc := mocks.Servicer{}
+			accountSvc.On("List", mock.MatchedBy(func(input *account.Account) bool {
+				if (input.ID != nil && tt.query.ID != nil && *input.ID == *tt.query.ID) || input.ID == tt.query.ID {
+					if tt.nextID != nil {
+						input.NextID = tt.nextID
+						input.Limit = ptr64(1)
+					}
+					return true
+				}
+				return false
+			})).Return(
+				tt.retAccounts, tt.retErr,
+			)
+			svcBldr.Config.WithService(&accountSvc)
+			_, err = svcBldr.Build()
+
+			assert.Nil(t, err)
+			if err == nil {
+				Services = svcBldr
+			}
+
+			GetAccounts(w, r)
+
+			resp := w.Result()
+			body, err := ioutil.ReadAll(resp.Body)
+
+			assert.Nil(t, err)
+			assert.Equal(t, tt.expResp.StatusCode, resp.StatusCode)
+			assert.Equal(t, tt.expResp.Body, string(body))
+			assert.Equal(t, tt.expLink, w.Header().Get("Link"))
+		})
 	}
+
 }