+++ pkg/data/accounts.go
@@ -9,8 +9,13 @@ import (
 	"github.com/aws/aws-sdk-go/service/dynamodb/expression"
 )
 
-// query for doing a query against dynamodb
-func (a *Account) query(query *account.Account, keyName string, index string) (*account.Accounts, error) {
+type queryScanOutput struct {
+	items            []map[string]*dynamodb.AttributeValue
+	lastEvaluatedKey map[string]*dynamodb.AttributeValue
+}
+
+// queryAccounts for doing a query against dynamodb
+func (a *Account) queryAccounts(query *account.Account, keyName string, index string) (*queryScanOutput, error) {
 	var expr expression.Expression
 	var bldr expression.Builder
 	var err error
@@ -27,7 +32,7 @@ func (a *Account) query(query *account.Account, keyName string, index string) (*
 		return nil, errors.NewInternalServer("unable to build query", err)
 	}
 
-	res, err = a.DynamoDB.Query(&dynamodb.QueryInput{
+	queryInput := &dynamodb.QueryInput{
 		TableName:                 aws.String(a.TableName),
 		IndexName:                 aws.String(index),
 		KeyConditionExpression:    expr.KeyCondition(),
@@ -35,8 +40,19 @@ func (a *Account) query(query *account.Account, keyName string, index string) (*
 		FilterExpression:          expr.Filter(),
 		ExpressionAttributeNames:  expr.Names(),
 		ExpressionAttributeValues: expr.Values(),
-	})
+	}
 
+	queryInput.SetLimit(*query.Limit)
+	if query.NextID != nil {
+		// Should be more dynamic
+		queryInput.SetExclusiveStartKey(map[string]*dynamodb.AttributeValue{
+			"Id": &dynamodb.AttributeValue{
+				S: query.NextID,
+			},
+		})
+	}
+
+	res, err = a.DynamoDB.Query(queryInput)
 	if err != nil {
 		return nil, errors.NewInternalServer(
 			"failed to query accounts",
@@ -44,16 +60,14 @@ func (a *Account) query(query *account.Account, keyName string, index string) (*
 		)
 	}
 
-	accounts := &account.Accounts{}
-	err = dynamodbattribute.UnmarshalListOfMaps(res.Items, accounts)
-	if err != nil {
-		return nil, errors.NewInternalServer("failed unmarshaling of accounts", err)
-	}
-	return accounts, nil
+	return &queryScanOutput{
+		items:            res.Items,
+		lastEvaluatedKey: res.LastEvaluatedKey,
+	}, nil
 }
 
-// scan for doing a scan against dynamodb
-func (a *Account) scan(query *account.Account) (*account.Accounts, error) {
+// scanAccounts for doing a scan against dynamodb
+func (a *Account) scanAccounts(query *account.Account) (*queryScanOutput, error) {
 	var expr expression.Expression
 	var err error
 	var res *dynamodb.ScanOutput
@@ -65,30 +79,66 @@ func (a *Account) scan(query *account.Account) (*account.Accounts, error) {
 			return nil, errors.NewInternalServer("unable to build query", err)
 		}
 	}
-	res, err = a.DynamoDB.Scan(&dynamodb.ScanInput{
+
+	scanInput := &dynamodb.ScanInput{
 		TableName:                 aws.String(a.TableName),
 		ConsistentRead:            aws.Bool(a.ConsistentRead),
 		FilterExpression:          expr.Filter(),
 		ExpressionAttributeNames:  expr.Names(),
 		ExpressionAttributeValues: expr.Values(),
-	})
+	}
+
+	scanInput.SetLimit(*query.Limit)
+	if query.NextID != nil {
+		// Should be more dynamic
+		scanInput.SetExclusiveStartKey(map[string]*dynamodb.AttributeValue{
+			"Id": &dynamodb.AttributeValue{
+				S: query.NextID,
+			},
+		})
+	}
+
+	res, err = a.DynamoDB.Scan(scanInput)
+
 	if err != nil {
 		return nil, errors.NewInternalServer("error getting accounts", err)
 	}
 
-	accounts := &account.Accounts{}
-	err = dynamodbattribute.UnmarshalListOfMaps(res.Items, accounts)
-	if err != nil {
-		return nil, errors.NewInternalServer("failed unmarshaling of accounts", err)
-	}
-	return accounts, err
+	return &queryScanOutput{
+		items:            res.Items,
+		lastEvaluatedKey: res.LastEvaluatedKey,
+	}, nil
 }
 
 // List Get a list of accounts
 func (a *Account) List(query *account.Account) (*account.Accounts, error) {
 
-	if query.Status != nil {
-		return a.query(query, "AccountStatus", "AccountStatus")
+	var outputs *queryScanOutput
+	var err error
+
+	if query.Limit == nil {
+		query.Limit = &a.Limit
 	}
-	return a.scan(query)
+
+	if query.Status != nil {
+		outputs, err = a.queryAccounts(query, "AccountStatus", "AccountStatus")
+	} else {
+		outputs, err = a.scanAccounts(query)
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	query.NextID = nil
+	for _, v := range outputs.lastEvaluatedKey {
+		query.NextID = v.S
+	}
+
+	accounts := &account.Accounts{}
+	err = dynamodbattribute.UnmarshalListOfMaps(outputs.items, accounts)
+	if err != nil {
+		return nil, errors.NewInternalServer("failed unmarshaling of accounts", err)
+	}
+
+	return accounts, nil
 }