+++ pkg/data/helpers.go
@@ -15,25 +15,26 @@ func getFiltersFromStruct(i interface{}, keyName *string) (*expression.KeyCondit
 	v := reflect.ValueOf(i).Elem()
 	for i := 0; i < v.NumField(); i++ {
 		dField := strings.Split(v.Type().Field(i).Tag.Get("dynamodbav"), ",")[0]
-		value := v.Field(i).Interface()
-		if !reflect.ValueOf(value).IsNil() {
-			switch v.Field(i).Kind() {
-			case reflect.Ptr:
-				dValue := reflect.Indirect(v.Field(i)).Interface()
-				if keyName != nil {
-					if dField == *keyName {
-						newFilter := expression.Key(dField).Equal(expression.Value(dValue))
-						kb = &newFilter
-						continue
+		if dField != "-" {
+			value := v.Field(i).Interface()
+			if !reflect.ValueOf(value).IsNil() {
+				switch v.Field(i).Kind() {
+				case reflect.Ptr:
+					dValue := reflect.Indirect(v.Field(i)).Interface()
+					if keyName != nil {
+						if dField == *keyName {
+							newFilter := expression.Key(dField).Equal(expression.Value(dValue))
+							kb = &newFilter
+							continue
+						}
+					}
+					if cb == nil {
+						newFilter := expression.Name(dField).Equal(expression.Value(dValue))
+						cb = &newFilter
+					} else {
+						*cb = cb.And(expression.Name(dField).Equal(expression.Value(dValue)))
 					}
 				}
-				if cb == nil {
-					newFilter := expression.Name(dField).Equal(expression.Value(dValue))
-					cb = &newFilter
-				} else {
-					*cb = cb.And(expression.Name(dField).Equal(expression.Value(dValue)))
-				}
-
 			}
 		}
 	}