+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add CloudWatch Dashboard for monitoring DCE
 - Support deleting a lease by ID at `GET /leases/{ID}` endpoint
+- Update `GET /accounts` to allow for querying with `adminRoleArn`, `principalRoleArn`, and `principalPolicyHash`
 
 ## v0.26.0